+++ src/web/src/Components.ts
@@ -1,8 +1,8 @@
 // Pages
 export { default as ProfilePage } from './Pages/ProfilePage';
 export { default as Page } from './Pages/Page';
-export { default as Valorant } from './Pages/game/Valorant';
-export { default as CS2 } from './Pages/game/CS2';
+export { default as Valorant } from './Pages/game/Valorant/Valorant';
+export { default as CS2 } from './Pages/game/CS2/CS2';
 export { default as Login } from './Pages/auth/Login';
 export { default as Register } from './Pages/auth/Register';
 export { default as Upload } from './Pages/upload/Upload';
@@ -24,22 +24,22 @@ export { default as AdminModifyUser } from './Pages/admin/AdminModifyUser';
 export { default as AdminModifyPost } from './Pages/admin/AdminModifyPost';
 export { default as PageNotFound } from './Pages/auth/PageNotFound';
 // Components
-export { default as Header } from './Components/Header';
-export { default as Game } from './Components/Game';
+export { default as Header } from './Components/global/Header';
+export { default as Game } from './Components/game/Game';
 export { default as SideNav } from './Components/sidebar/SideNav';
 export { default as SideNavItems } from './Components/sidebar/SideNavItems';
-export { default as Footer } from './Components/Footer';
+export { default as Footer } from './Components/global/Footer';
 export { default as SideNavWrapper } from './Components/sidebar/SideNavWrapper';
-export { default as Searchbar } from './Components/Searchbar';
+export { default as Searchbar } from './Components/game/Searchbar';
 export { default as RequireAuth } from './Components/auth/RequireAuth';
-export { default as Dropzone } from './Components/Dropzone';
-export { default as Carousel } from './Components/Carousel';
-export { default as Loading } from './Components/Loading';
-export { default as Posts } from './Components/Posts';
-export { default as ProfileEdit } from './Components/ProfileEdit';
-export { default as AgentSelector } from './Components/AgentSelector';
-export { default as ValorantPopup } from './Components/ValorantPopup';
-export { default as CookiesPopup } from './Components/CookiesPopup';
+export { default as Dropzone } from './Components/upload/Dropzone';
+export { default as Carousel } from './Components/game/Carousel';
+export { default as Loading } from './Components/global/Loading';
+export { default as Posts } from './Components/post/Posts';
+export { default as ProfileEdit } from './Components/profile/ProfileEdit';
+export { default as AgentSelector } from './Components/game/valorant/AgentSelector';
+export { default as ValorantPopup } from './Components/game/valorant/ValorantPopup';
+export { default as CookiesPopup } from './Components/auth/CookiesPopup';
 export { default as ValorantMode } from './Components/upload/ValorantMode';
 export { default as CS2Mode } from './Components/upload/CS2Mode';
-export { default as Dot } from './Components/Dot';
+export { default as Dot } from './Components/upload/Dot';
