+++ src/web/src/Components/game/Carousel.tsx
