+++ src/web/src/Components/upload/CS2Mode.tsx
@@ -94,7 +94,14 @@ const CS2Mode: React.FC<CS2ModeProps> = ({ state, dispatch }) => {
             context?.drawImage(img, 0, 0, img.width, img.height);
          }
       }
-   }, [mapImage, clickPosition, setClickPosition, hoverPosition]);
+
+      if (placedDot) {
+         dispatch({
+            type: 'setLineupPositionCoords',
+            payload: { x: placedDot.x, y: placedDot.y },
+         });
+      }
+   }, [mapImage, clickPosition, setClickPosition, hoverPosition, dispatch, placedDot]);
 
    const handleMouseMove = (
       e: React.MouseEvent<HTMLCanvasElement, MouseEvent>,
@@ -163,16 +170,10 @@ const CS2Mode: React.FC<CS2ModeProps> = ({ state, dispatch }) => {
                   type: 'setLineupLocationCoords',
                   payload: { x: selectedDot.x, y: selectedDot.y, name: selectedDot.name },
                });
-               if (placedDot) {
-                  dispatch({
-                     type: 'setLineupPositionCoords',
-                     payload: { x: placedDot.x, y: placedDot.y },
-                  });
-               }
             }
          }
 
-         setClickPosition({ x, y});
+         setClickPosition({ x, y });
       }
    };
